Fix authorize errors

+++ src/backend/SportLink.Backend/SportLink.API/Services/Organization/OrganizationService.cs
@@ -189,7 +189,13 @@ namespace SportLink.API.Services.Organization
 
         public async Task<ActionResult<OrganizationDetailedDto>> UpdateProfile(int id, OrganizationDetailedDto organizationDetailedDto)
         {
+            var ownerId = _httpContextAccessor.HttpContext?.User?.Claims.FirstOrDefault(x => x.Type == ClaimTypes.NameIdentifier)?.Value;
             var organization = await _context.Organizations.FindAsync(id);
+            if (organization is null || organization.OwnerId != int.Parse(ownerId!))
+            {
+                return new BadRequestResult();
+            }
+
             var existingSocialNetworks = await _context.SocialNetworks.Where(x => x.OrganizationId == id).ToListAsync();
             if (organization is not null && organization.VerificationStatus == VerificationStatusEnum.Accepted)
             {
@@ -239,7 +245,12 @@ namespace SportLink.API.Services.Organization
 
         public async Task<bool> DeleteOrganization(int id)
         {
+            var ownerId = _httpContextAccessor.HttpContext?.User?.Claims.FirstOrDefault(x => x.Type == ClaimTypes.NameIdentifier)?.Value;
             var organization = await _context.Organizations.FindAsync(id);
+            if (organization is null || organization.OwnerId != int.Parse(ownerId!))
+            {
+                return false;
+            }
             if (organization is not null && organization.VerificationStatus == VerificationStatusEnum.Accepted)
             {
                 _context.Organizations.Remove(organization);
